requirements.txt
flask==3.0.0
google-genai>=1.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
pypdf>=4.0.0
pydantic>=2.0.0
pandas>=2.0.0
groq>=0.4.0
